Change file headers, varible names and file names
To have a more consistent file layout, a uniform header template is
applied to all *.m-files.

In build_sys_matrix.m the output variables Ksys and Msys are renamed to
KSys and MSys to be more consistent with the other functions and script
files.

The main script file frame-fem.m is renamed to frame_fem.m to avoid
problems with MATLAB not being able to run the script because of the
dash inside the filename.