Add finite-element system class
Addition of the class c_sys_fem, which represents the discretized frame
system including all nodes, elements and the system mass and stiffness
matrices.

The class assembles the system matrices by individually adding elements
to the system.

The main script file is altered to use the new class instead of the old
build_sys_matrix-function.