Add frame defintion class
A new class c_frame_def is introduced, which creates the nodes vector
for the discretized system of a twodimensional frame structure. The
class needs a list of "parent" nodes where the "parent" beams, specified
by another list connecting certain nodes with beams of different
properties, are discretized by a different amount of elements.

The main script, frame_fem, is altered to account for these changes and
let the new class create the nodes vector neccessary for the discretized
fem-system and system matrices.