Add infinite boundary condition assignment
To account for infinite element boundary conditions (wave-transmissive)
using specially tuned spring-damper-elements (WANG and LAI, 1999),
several methods and properties are introduced to the classes c_frame_def
and c_sys_fem.

Note that these BCs aren't useable for harmonic analyses yet and
furthermore only work in the global z-direction and not for nodes with
multiple beams connected to them.