Add infinite beam element to the left side
To allow energy to flow out of the domain on the left hand side of the
double-frame, another infinite beam element is attached just in front of
the harmonic force excitation point.

Additionally the transmission coefficient of the double-frame is now
calculated using the ratio of outgoing power on the right hand side and
overall incoming power.